+++ src/main/java/roomescape/payment/api/PaymentClient.java
@@ -1,16 +1,16 @@
-package roomescape.payment.config;
+package roomescape.payment.api;
 
 import java.util.Base64;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import roomescape.payment.config.PaymentClientResponseErrorHandler;
 import roomescape.payment.dto.PaymentRequest;
 import roomescape.payment.dto.PaymentResponse;
 
@@ -21,7 +21,6 @@ public class PaymentClient {
     private static final String AUTH_METHOD = "Basic ";
     private static final Logger log = LoggerFactory.getLogger(PaymentClient.class);
     private final String encodedSecretKey;
-    private final String paymentUrl;
     private final RestClient restClient;
     private final ObjectMapper objectMapper;
 
@@ -31,7 +30,6 @@ public class PaymentClient {
                          ObjectMapper objectMapper) {
         this.encodedSecretKey = Base64.getEncoder()
                 .encodeToString((secretKey + DELIMITER).getBytes());
-        this.paymentUrl = paymentUrl;
         this.restClient = restClientBuilder
                 .baseUrl(paymentUrl)
                 .build();
@@ -47,7 +45,7 @@ public class PaymentClient {
                 .retrieve()
                 .onStatus(new PaymentClientResponseErrorHandler(objectMapper))
                 .body(PaymentResponse.class);
-        log.info("URI: {}, Method: {}, Body:{} ", uri, HttpMethod.POST, paymentResponse);
+        log.info("URI: {}, Method: {}, Body:{} ", uri, "POST", paymentResponse);
         return paymentResponse;
     }
 }
