+++ src/test/java/roomescape/integration/controller/ReservationControllerTest.java
@@ -33,8 +33,7 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import roomescape.exception.PaymentException;
 import roomescape.exception.response.PaymentExceptionResponse;
-import roomescape.exception.type.PaymentExceptionType;
-import roomescape.payment.config.PaymentClient;
+import roomescape.payment.api.PaymentClient;
 import roomescape.auth.domain.Role;
 import roomescape.payment.dto.PaymentRequest;
 import roomescape.reservation.dto.ReservationResponse;
